refactor: reduce responsibilities of dashboard component (pass them on to modal & table)
